move ActivityDiagnosticTask into a separate header
